:hammer: refactor control uri-handled tasks
Put each control task in a different file. Since there aren't plans to
be too many they are handled in the same level as server.
A very unpleasant global "URI_TASK" is shared between those files to
keep track of the task we are in when handling the websocket callbacks.

Right now the tasks are dummy and only return one string via websocket
with information about the task. In the future, this tasks should read
the sensors and write to the actuators. Then write this info to the
database via mutexes.