:sparkles: add driver for quadrature encoder
Driver for quadrature encoder with ESP8266.

This driver works with the usual rotary quadrature encoder, typically
found in printers. It consists in two optocouplers, which, in prescence
of rotation produce square signals in a 90 degree phase.

The purpose of this driver is to read angular position, therefore no
timer is needed.

The driver is interrupt based, trying to minimize in-interrupt time.