:wrench: fix header in data bug
From time to time, a wild part of the header appeared in the middle of
the data received in the websocket. The problem was that the header
array is the same as the data array. And when data is received, the
array is just overwritten, not flush.

Altough "always flushing the array" would be a nice-to-have in the
websockets ESP library, it can be workarrounded very easily. A null
character is added at the data_len position of the array. Producing a
"string end" for every (?) use of said array in this code.