:wrench: fix uri unrecognized bug
The test_task was not thought to be run multiple times (other tasks may
 have the same problem). And some "expected time to live" is given when
it is started. Then, when it ends, it assigns the global URI_TASK to an
URI_UNDEFINED state.

When a task like this runs once, there is no problems, as there is no
problem in tests. But when runs multiple times, it will:

 * If a new handshake is done, it will cause overlapping tasks,
consuming memory.
 * If we don't run out of memory first (comms hanging) we will most
likely have one of this tasks finishing and reseting the URI_TASK
status to URI_UNDEF. Hence, uris pass as unrecognized.

To fix this, the task (as in classic controller) runs until the comms
are closed. This can happen if another connection is open (new uri), so
will avoid overlapping, and also prevents the task from writing
URI_UNDEF to the URI_TASK global, since does it only when closing.