:wrench: workarround pwm turning off
When the PWM signal is seen in the oscilloscope, it shows how the PWM
problems happen when the duty is changed (and the PWM is reset). This
abrupt "cut" can happen during the active duty cycle. If it happens
after a duty greater than 1 ms, the motor driver will interpret this as
a new throttling command different from the intended one.

This workarround updates the PWM only when the duty value is different
from the one set, decreasing the chances for this issue to happen.