:wrench: avoid drowning the idle task
The low priority 'httpd_task' went in a busy wait after creating the
other tasks and callbacks. Since the telemetry and actuators tasks had
more priority, this wasn't much of a problem.

The negative connotation of this approach is the lack of time available
for the idle task (lowest priority of all by default). The idle task is
in charge of freeing the memory that was allocated for the now-deleted
tasks [1].

Instead of a busy for, add a vTaskDelay() that will hand control to the
scheduler.

[1] https://www.freertos.org/a00126.html